+++ py/desispec/test/test_ql_qa.py
@@ -91,6 +91,7 @@ class TestQL(unittest.TestCase):
         hdr['EXPID'] = 1
         hdr['PROGRAM'] = 'dark'
         hdr['FLAVOR'] = 'science'
+        hdr['EXPTIME'] = 100.0
         
         rawimage = np.zeros((2*ny, 2*nx+2*noverscan))
         offset = {'1':100.0, '2':100.5, '3':50.3, '4':200.4}
@@ -170,7 +171,7 @@ class TestQL(unittest.TestCase):
         ivar=np.ones_like(flux)
         resolution_data=np.ones((nspec,13,nwave))
         self.frame=desispec.frame.Frame(wave,flux,ivar,resolution_data=resolution_data,fibermap=self.fibermap)
-        self.frame.meta = dict(CAMERA=self.camera,PROGRAM='dark',FLAVOR='science',NIGHT=self.night, EXPID=self.expid,CCDSEC1=self.ccdsec1,CCDSEC2=self.ccdsec2,CCDSEC3=self.ccdsec3,CCDSEC4=self.ccdsec4)
+        self.frame.meta = dict(CAMERA=self.camera,PROGRAM='dark',FLAVOR='science',NIGHT=self.night,EXPID=self.expid,EXPTIME=100,CCDSEC1=self.ccdsec1,CCDSEC2=self.ccdsec2,CCDSEC3=self.ccdsec3,CCDSEC4=self.ccdsec4)
         desispec.io.write_frame(self.framefile, self.frame)
 
         #- make a skymodel
@@ -207,9 +208,6 @@ class TestQL(unittest.TestCase):
         counts1=qalib.countpix(pix,nsig=3) #- counts above 3 sigma
         counts2=qalib.countpix(pix,nsig=4) #- counts above 4 sigma
         self.assertLess(counts2,counts1)
-        counts3=qalib.countpix(pix,ncounts=200)
-        counts4=qalib.countpix(pix,ncounts=250)
-        self.assertLess(counts4,counts3)
 
     def test_sky_resid(self):
         import copy
@@ -368,13 +366,14 @@ class TestQL(unittest.TestCase):
         #- test if amp QAs exist
         qargs["amps"] = True
         resl2=qa(inp,**qargs)
-        self.assertTrue(len(resl2['METRICS']['NPIX3SIG_AMP'])==4)
+        self.assertTrue(len(resl2['METRICS']['NPIX_LOW_AMP'])==4)
 
     def testCountSpectralBins(self):
         qa=QA.CountSpectralBins('countbins',self.config)
         inp=self.frame
         qargs={}
         qargs["PSFFile"]=self.psf
+        qargs["FiberMap"]=self.fibermap
         qargs["camera"]=self.camera
         qargs["expid"]=self.expid
         qargs["amps"]=True
@@ -390,6 +389,7 @@ class TestQL(unittest.TestCase):
         qa=QA.Sky_Continuum('skycont',self.config)
         inp=self.frame
         qargs={}
+        qargs["FiberMap"]=self.fibermap
         qargs["camera"]=self.camera
         qargs["expid"]=self.expid
         qargs["amps"]=False
@@ -407,6 +407,7 @@ class TestQL(unittest.TestCase):
         qa=QA.Sky_Peaks('skypeaks',self.config)
         inp=self.frame
         qargs={}
+        qargs["FiberMap"]=self.fibermap
         qargs["camera"]=self.camera
         qargs["expid"]=self.expid
         qargs["amps"]=True
@@ -421,6 +422,7 @@ class TestQL(unittest.TestCase):
         inp=self.frame
         qargs={}
         qargs["PSFFile"]=self.psf
+        qargs["FiberMap"]=self.fibermap
         qargs["camera"]=self.camera
         qargs["expid"]=self.expid
         qargs["amps"]=False
@@ -441,6 +443,7 @@ class TestQL(unittest.TestCase):
         sky=self.skymodel
         qargs={}
         qargs["PSFFile"]=self.psf
+        qargs["FiberMap"]=self.fibermap
         qargs["camera"]=self.camera
         qargs["expid"]=self.expid
         qargs["amps"]=True
@@ -461,6 +464,7 @@ class TestQL(unittest.TestCase):
         inp=self.frame
         qargs={}
         qargs["PSFFile"]=self.psf
+        qargs["FiberMap"]=self.fibermap
         qargs["camera"]=self.camera
         qargs["expid"]=self.expid
         qargs["amps"]=True