+++ doc/changes.rst
@@ -6,6 +6,8 @@ desispec Change Log
 -------------------
 
 * Small fixes to desi_qa_prod and qa_prod
+* Fixes integration tests for desisim newexp refactor
+* Removes spectra grouping by brick; nside=64 healpix grouping default
 
 0.15.2 (2017-07-12)
 -------------------